+++ DEMO/mySubscriber/MySubscriber.h
@@ -4,7 +4,7 @@
 #include <nrt/Core/Blackboard/Module.H>
 
 NRT_DECLARE_MESSAGESUBSCRIBER_PORT(SomeTextPort,   nrt::Message<std::string>, void, "Some text to print");
-NRT_DECLARE_MESSAGESUBSCRIBER_PORT(SomeNumberPort, nrt::Message<float>,       void, "Some number to print");
+NRT_DECLARE_MESSAGESUBSCRIBER_PORT(SomeNumberPort, nrt::Message<double>,       void, "Some number to print");
 
 class MySubscriber : public nrt::Module,
                      public nrt::MessageSubscriber<SomeTextPort, SomeNumberPort>
